merge

+++ Connections/MSPBSL_Connection_v1_4x.cpp
@@ -102,7 +102,7 @@ uint16_t MSPBSL_Connection_v1_4x::TX_BSL_Version(string& versionString)
 * \return the value returned by the connected BSL, or underlying connection layers
 ******************************************************************************/
 
-uint16_t MSPBSL_Connection_v1_4x::eraseCheck( uint16_t startAddr, uint32_t numBytes )
+uint16_t MSPBSL_Connection_v1_4x::eraseCheck( uint32_t startAddr, uint32_t numBytes )
 {
 	uint16_t retValue = 0;
